+++ src/icons/b-icon/test/index.js
@@ -30,12 +30,12 @@ module.exports = (page) => {
 				target = await init({value: 'foo'});
 
 			expect(clrfx(await target.evaluate((ctx) => ctx.$el.innerHTML)))
-				.toBe('<use data-update-on-id="id"></use><use xlink:href="/p-v4-components-demo.html#foo" data-tmp=""></use>');
+				.toBe('<use data-update-on-id="id"></use><use xlink:href="#foo" data-tmp=""></use>');
 
 			await Promise.resolve('next tick');
 
 			expect(clrfx(await target.evaluate((ctx) => ctx.$el.innerHTML)))
-				.toBe('<use data-update-on-id="id"></use><use xlink:href="/p-v4-components-demo.html#foo" data-tmp=""></use>');
+				.toBe('<use data-update-on-id="id"></use><use xlink:href="#foo" data-tmp=""></use>');
 		});
 	});